hibajavitasok

+++ UserApi/Controllers/UserController.cs
@@ -1,5 +1,8 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using UserApi.Models;
 
 namespace UserApi.Controllers
@@ -9,34 +12,15 @@ namespace UserApi.Controllers
     public class UserController : ControllerBase
     {
         [HttpGet]
-        public ActionResult<List<User>> Get()
+        public ActionResult<List<User>> GetAll()
         {
             using (var context = new UserDbContext())
             {
-                return StatusCode(201, NewMethod(context));
+                var users = context.Users.ToList();
+                return Ok(users);
             }
         }
 
-        private static object NewMethod(UserDbContext context)
-        {
-            return context.newUser.ToList();
-        }
-
-        private ActionResult<List<User>> StatusCode(int v, object value)
-        {
-            throw new NotImplementedException();
-        }
-
-        [HttpGet]
-        public ActionResult<User> Post(CreateUserDto user)
-        {
-            using (var context = new UserDbContext())
-            {
-                return Ok();
-
-            }
-        }
-        #region
         [HttpGet("{id}")]
         public ActionResult<User> GetById(Guid id)
         {
@@ -51,27 +35,14 @@ namespace UserApi.Controllers
                 return Ok(user);
             }
         }
-        #endregion
-        [HttpPut("{id}")]
-        public ActionResult<User> Put(Guid id, [FromBody] CreateUserDto userDto)
-        {
-
-
-            using (var context = new UserDbContext())
-            {
-                var existingUser = context.Users.Find(id);
-               
-                existingUser.Name = userDto.Name;
-
-                context.SaveChanges();
-                return Ok(existingUser);
-            }
-        }
 
         [HttpPost]
         public ActionResult<User> Post([FromBody] CreateUserDto userDto)
         {
-
+            if (userDto == null)
+            {
+                return BadRequest("A felhasználói adatok szükségesek.");
+            }
 
             using (var context = new UserDbContext())
             {
@@ -85,19 +56,48 @@ namespace UserApi.Controllers
                 context.Users.Add(user);
                 context.SaveChanges();
 
-                return CreatedAtAction(nameof(Get), new { id = user.Id }, user);
+                return CreatedAtAction(nameof(GetById), new { id = user.Id }, user);
+            }
+        }
+
+        [HttpPut("{id}")]
+        public ActionResult<User> Put(Guid id, [FromBody] CreateUserDto userDto)
+        {
+            if (userDto == null)
+            {
+                return BadRequest("A felhasználói adatok szükségesek.");
+            }
+
+            using (var context = new UserDbContext())
+            {
+                var existingUser = context.Users.Find(id);
+                if (existingUser == null)
+                {
+                    return NotFound();
+                }
+
+                existingUser.Name = userDto.Name;
+                existingUser.Email = userDto.Email;
+
+                context.SaveChanges();
+                return Ok(existingUser);
             }
         }
 
         [HttpDelete("{id}")]
-        public ActionResult<object> Delete(Guid id)
+        public ActionResult Delete(Guid id)
         {
             using (var context = new UserDbContext())
             {
                 var user = context.Users.Find(id);
+                if (user == null)
+                {
+                    return NotFound();
+                }
+
                 context.Users.Remove(user);
                 context.SaveChanges();
-                return Ok(new { message = "Törölve!" });
+                return Ok(new { message = "A felhasználó sikeresen törölve!" });
             }
         }
     }
